requirements.txt
python3-scapy
python3-gpiozero